+++ tests/enhydris_api_client/test_station.py
@@ -1,10 +1,94 @@
 from unittest import TestCase
+from unittest.mock import call
 
 import requests
 
-from enhydris_api_client import EnhydrisApiClient
+from enhydris_api_client import EnhydrisApiClient, MalformedResponseError
 
 from . import mock_session
+
+
+class ListStationsSinglePageTestCase(TestCase):
+    @mock_session()
+    def setUp(self, m):
+        m.return_value.get.return_value.json.return_value = {
+            "count": 2,
+            "next": None,
+            "previous": None,
+            "results": [{"name": "Hobbiton"}, {"name": "Rivendell"}],
+        }
+        self.mock_session = m
+        client = EnhydrisApiClient("https://mydomain.com")
+        self.result = client.list_stations()
+
+    def test_makes_request(self):
+        next(self.result)  # Ensure the request is actually made
+        m = self.mock_session
+        m.return_value.get.assert_called_once_with("https://mydomain.com/api/stations/")
+
+    def test_result(self):
+        self.assertEqual(
+            list(self.result),
+            [{"name": "Hobbiton"}, {"name": "Rivendell"}],
+        )
+
+
+class ListStationsMultiPageTestCase(TestCase):
+    @mock_session()
+    def setUp(self, m):
+        m.return_value.get.return_value.json.side_effect = [
+            {
+                "count": 3,
+                "next": "https://mydomain.com/api/stations/?page=2",
+                "previous": None,
+                "results": [{"name": "Hobbiton"}, {"name": "Rivendell"}],
+            },
+            {
+                "count": 3,
+                "next": None,
+                "previous": "https://mydomain.com/api/stations/",
+                "results": [{"name": "Mordor"}],
+            },
+        ]
+        self.mock_session = m
+        client = EnhydrisApiClient("https://mydomain.com")
+        self.result = client.list_stations()
+
+    def test_requests(self):
+        list(self.result)  # Ensure all requests are made
+        self.assertEqual(
+            self.mock_session.return_value.get.call_args_list,
+            [
+                call("https://mydomain.com/api/stations/"),
+                call("https://mydomain.com/api/stations/?page=2"),
+            ],
+        )
+
+    def test_result(self):
+        self.assertEqual(
+            list(self.result),
+            [{"name": "Hobbiton"}, {"name": "Rivendell"}, {"name": "Mordor"}],
+        )
+
+
+class ListStationsErrorTestCase(TestCase):
+    @mock_session(**{"get.return_value.status_code": 500})
+    def test_raises_exception_on_error(self, m):
+        client = EnhydrisApiClient("https://mydomain.com")
+        with self.assertRaises(requests.HTTPError):
+            next(client.list_stations())
+
+    @mock_session(**{"get.return_value.json.return_value": "not a dict"})
+    def test_raises_exception_on_non_json_response(self, m):
+        client = EnhydrisApiClient("https://mydomain.com")
+        with self.assertRaises(MalformedResponseError):
+            next(client.list_stations())
+
+    @mock_session(**{"get.return_value.json.return_value": {"no": "expected"}})
+    def test_raises_exception_on_unexpected_json(self, m):
+        client = EnhydrisApiClient("https://mydomain.com")
+        with self.assertRaises(MalformedResponseError):
+            next(client.list_stations())
 
 
 class GetStationTestCase(TestCase):