+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+2.3.0 (2025-09-13)
+==================
+
+* Compatible with Python 3.13.
+* Added ``enhydris_api_client.EnhydrisApiClient.list_stations()``.
+
 2.2.3 (2025-03-26)
 ==================